requirements.txt
annotated-types==0.6.0
anthropic==0.19.2
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.12.0
packaging==24.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
tiktoken==0.6.0
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
